Project structure updated to only one view class.
