--- src/main/java/Tax.java
@@ -1,3 +1,0 @@
-public enum Tax {
-
-}
